requirements.txt
langchain-core~=0.3.15
pandas~=2.2.2
python-dotenv~=1.0.1
openai~=1.54.3
openpyxl~=3.1.5
promptflow~=1.16.1
tenacity~=9.0.0
diskcache~=5.6.3
azure-ai-documentintelligence==1.0.0b2
azure-ai-formrecognizer==3.3
langchain-text-splitters==0.3.2
langchain==0.3.8
tiktoken==0.8.0
pydantic==2.8.2
aiofiles==24.1.0
azure-search-documents==11.6.0b4
